Add CLI 'erase' command to wipe NDEF content of the next tag
  Implements a one-shot erase flow driven by the CLI: typing `erase`
  arms a flag; on the next tag activation the IoT reader writes an
  empty NDEF message and clears the flag. A second `erase` disarms.

    - T2T  : ptxNDEF_T2TOp Open / CheckMessage / WriteMessage(len=0).
    - T4T  : raw-APDU helper ptxIoTRdInt_EraseType4NDEF
             (SELECT NDEF App -> SELECT/READ CC -> SELECT NDEF file
             -> UPDATE BINARY NLEN=0). The full T4T Op layer (~3 KB)
             does not fit on the RA2E3.
    - T3T/T5T are intentionally omitted so their Op code is dropped
      by --gc-sections, keeping the image inside the 63 KB flash
      budget.

  To free flash for the erase path, two large decoders in
  ptx_IOT_RD_Main.c were replaced with lightweight fallbacks:
    - PrintNDEFMsg: the full NFC Forum decoder (URI prefix table,
      Text / Smart Poster / Wi-Fi / Bluetooth / MIME / External
      dispatch, ~1.7 KB) is replaced by a raw hex dump. The
      TypeEq / StartsWith / WscFind / PrintWifi / PrintBt helpers
      are marked `static inline` so the linker drops them.
    - PrintCardInfo: the ISO-DEP EMV path (PPSE / AID / GPO / AFL /
      READ RECORD) is replaced by a one-line "Non-NDEF ISO-DEP card"
      report.

  CLI / UX:
    - Line buffer 80 -> 120, prompt "> " -> "$ ".
    - Renamed ledon / ledoff -> lon / loff.
    - cli_prompt is no longer static; added UserCli_ArmEraseNext /
      UserCli_IsEraseArmed / UserCli_ClearEraseArmed used by the
      NFC main loop.

  Minor cleanups:
    - ptxCOMMON.c: cast strnlen() result to size_t for
      UserUartLog_Write (silences a sign/width warning).
    - ptx_IOT_READER.c, user_uart_log.c: add forward declarations to
      silence implicit-function-declaration warnings.
    - user_cli.h: include <stdint.h> for uint8_t in the public API.

+++ src/STACK/COMPS/COMMON/ptxCOMMON.c
@@ -103,7 +103,7 @@ void ptxCommon_PrintF(const char *format, ...)
 
 #ifdef ENABLE_PRINTF_OPTION_UART
     /* Mirror the formatted line to g_uart0 (TXD on P1_09). */
-    UserUartLog_Write((const uint8_t *)buffer, strnlen(buffer, max_len));
+    UserUartLog_Write((const uint8_t *)buffer, (size_t )strnlen(buffer, max_len));
 #endif
 
     va_end(argptr);
